Monitor sized styling
To cover landscape oriented monitors with resolution widths between 1200px and 1920px, responsive styling has been applied to the website header, website footer, and landing page.

The text, images, and/or element containers have been adjusted to change in accordance with the widths of the screens to retain similar layout for UX consistency.